Fixing Edit Sales Bug and Updating GUI
A new form was added which has replaced the Main form, this allows you to choose if you want to view the data base or generate a report. The form that used to be called main is now called database. The generate report portion of the new GUI is not yet implemented, i will work on this tommorow.